Make sha256 check command working on both Linux and Darwin
Makefile uses sha256sum command on both Linux and Darwin platforms to
check the integrity of the build. The problem is macOS does not ship the
sha sum utility with the same name. This commit uses the right utility
name for macOS platforms whilst preserving the original command for
usage on Linux OS.