Control state change triggers via templating
There was an issue with Terraform here that I couldn't get to the bottom
of.

There are two cases

1) I do not declare the binary as a dependency for the instance pool
instance and rely on the terraform default which should work.
2) I mark the binary as an explicit dependency with depends_on.

In case 1, when the binary changes (meaning the key changes as it's the
name) only the s3 object is updated, not the launch configs

In case 2, the launch configs are changed even if the binary is not.

I have spent time looking into the issue but the generated LCs are the
exact same so I'm unsure as to what's going on.

I have opted to avoid the terraform resource dependencies all together
in this commit.