Change resource so that we can run a clean terraform destroy
Some outputs were giving errors that it couldn't find an attribute of a resource when running terraform destroy. We had to implement a little hack/workaround to fix this. We add a count of 1 to the resource where the output gave an error. In the output we used the concat and element interpolation functions to fix the ouput. This is based on https://github.com/hashicorp/terraform/issues/16681#issuecomment-345105956 and we see this error since 0.11 (More information on that https://www.terraform.io/upgrade-guides/0-11.html#referencing-attributes-from-resources-with-count-0)