Remove unneeded input for Jenkins instance size
We can remove this variable both from inputs.tf.template and modules.tf.template because we already template it in from instance_pool_variables.template. In iinstance_pool_variables.template we use GO templating to set the correct instance size. This change fixes that we always brought up Jenkins as an t2.large instead of the size we intended.
Before this change we always installed Jenkins as