Create an SRV record for the etcd instance pool
When templating out the terraform config for the etcd stateful instance
pool, this will also create an SRV record in Route53 with a record for
each of the individual A records for each node.

This will eventually allow us to simplify the prometheus scrape config
the 'knowledge' of the number of nodes and their hostnames remains in
terraform only.

dig output:

```
[centos@charlie-cluster|master|ip-10-99-62-26 ~]$ dig etcd-svc.charlie-cluster.tarmak.local. SRV

; <<>> DiG 9.9.4-RedHat-9.9.4-61.el7 <<>> etcd-svc.charlie-cluster.tarmak.local. SRV
;; global options: +cmd
;; Got answer:
;; ->>HEADER<<- opcode: QUERY, status: NOERROR, id: 13292
;; flags: qr rd ra; QUERY: 1, ANSWER: 3, AUTHORITY: 0, ADDITIONAL: 1

;; OPT PSEUDOSECTION:
; EDNS: version: 0, flags:; udp: 4096
;; QUESTION SECTION:
;etcd-svc.charlie-cluster.tarmak.local. IN SRV

;; ANSWER SECTION:
etcd-svc.charlie-cluster.tarmak.local. 60 IN SRV 1 10 9115 etcd-2.charlie-cluster.
etcd-svc.charlie-cluster.tarmak.local. 60 IN SRV 1 10 9115 etcd-3.charlie-cluster.
etcd-svc.charlie-cluster.tarmak.local. 60 IN SRV 1 10 9115 etcd-1.charlie-cluster.

;; Query time: 3 msec
;; SERVER: 10.99.0.2#53(10.99.0.2)
;; WHEN: Tue Jun 26 15:32:42 UTC 2018
;; MSG SIZE  rcvd: 192
```