Add separate Packer builder for Kubernetes workers
We want to decrease the provision time when a new Kubernetes worker nodes comes up. We do this by creating a separate Packer build for the Kubernetes workers. This will run the Puppet code during the Packer run and install and configure almost everything. On initial boot we will only need to run Puppet to configure the cluster specific settings.

Signed-off-by: Mattias Gees <mattias.gees@gmail.com>