Prepare hub before terraforming other clusters
We had an issue (https://github.com/jetstack/tarmak/issues/377) where we
needed to plan the hub just to get the local files in place before
running operations against other clusters.

This 'Prepares' the hub cluster before verifying it which puts the
required files in place.

If the hub has not been applied, then trying to terraform another
cluster in the env will still fail.

I've also added a comment explaining this function a little more as it's
caused me some confusion in the past.