Fix verify of multicluster
When applying the kubernetes part in a multi-cluster, you get errors in the verify step of the verification of the hub.
This is because the variables are not correctly propagated and we ended up with a terraform state configuration of hub that was pointing to the state of the kubernetes part. This small fix fixes that.

Signed-off-by: Mattias Gees <mattias.gees@gmail.com>