Adds `plan --plan-file store` and `apply plan-file-location`
The plan-file-store flag sets the output location of the plan executable
file. The apply-file-location sets the file to be used when applying.
Both default to the current default path. If a plan-file-location has
been set during an apply, the normal plan that is made before the apply
is omitted, only the apply will run.

Signed-off-by: JoshVanL <vleeuwenjoshua@gmail.com>