Fix upper case of S in elasticsearch
The S was uppercased in the word elasticsearch. This caused problems in puppet as it needed to be lower case (https://github.com/jetstack/tarmak/blob/1598b6781a7fbd3c4554dcbd5864d90ef7b20d66/puppet/modules/fluent_bit/manifests/output.pp#L10). I have chosen to not change it in the puppet code, because elasticsearch is one word and the S is not written with an upper case in the official docs.
This has no impact on any existing config. I tested yaml config with both elasticsearch and elasticSearch.

Signed-off-by: Mattias Gees <mattias.gees@gmail.com>