Fix bug where we give false errors on instance availability
This fixes a bug where we give an error with Tarmak that an instance is not available while it is.
Becuase AWS doesn't provide an API endpoint to check instance availability, we have to use a hacky way to get it. That is why we use the reserved instance offering API.
Tarmak used to check every AZ, but this is not needed for the on-demand instances and region is enough. Now we only check if we get items back from AWS and continue.

Signed-off-by: Mattias Gees <mattias.gees@gmail.com>